move correlator to an internal package
