+++ state/kafka-sender/src/test/java/com/expediagroup/streamplatform/streamregistry/state/kafka/KafkaEventSenderTest.java
@@ -38,7 +38,7 @@ import org.mockito.Captor;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
-import com.expediagroup.streamplatform.streamregistry.state.EventCorrelator;
+import com.expediagroup.streamplatform.streamregistry.state.internal.EventCorrelator;
 import com.expediagroup.streamplatform.streamregistry.state.avro.AvroConverter;
 import com.expediagroup.streamplatform.streamregistry.state.avro.AvroEvent;
 import com.expediagroup.streamplatform.streamregistry.state.avro.AvroKey;
